feature: Full-text search of source for translation usage
You must have either `rg` or `rg.exe` in your path for this feature
to work. You must also pass an otherwise-optional list of source
directories to the tool when invoking it. If you do, the list of
unused translations has the ones used in source code removed. This
search right now is pretty brute-force and doesn't filter out
anything like documentation or json files, so the feature still needs
polish before I ship it.