feat: streamlined command-line options
We now deduce everything we can from a single passed-in mod
directory, searching for an `Interface` directory to identify
potential data dirs. This will not handle the case of a repo
with source for several data directories, perhaps in a fomod
archive structure, but it does handle the single-candidate
case. The ripgrep step, if rg is available, skips translation
files and all jsons when looking for uses in source code.

Added `walkdir` to the deps, but haven't used it everywhere yet.